Corrected sidecar invalid usage of int for block numbers.

+++ DiscImageChef.Core/Devices/Dumping/SBC.cs
@@ -708,7 +708,7 @@ namespace DiscImageChef.Core.Devices.Dumping
                 sidecar.OpticalDisc[0].DumpHardwareArray[0].Extents[0] = new ExtentType
                 {
                     Start = 0,
-                    End = (int)(blocks - 1)
+                    End = blocks - 1
                 };
                 sidecar.OpticalDisc[0].DumpHardwareArray[0].Manufacturer = dev.Manufacturer;
                 sidecar.OpticalDisc[0].DumpHardwareArray[0].Model = dev.Model;
@@ -831,7 +831,7 @@ namespace DiscImageChef.Core.Devices.Dumping
                     sidecar.BlockMedia[0].DumpHardwareArray[0].Extents[0] = new ExtentType
                     {
                         Start = 0,
-                        End = (int)(blocks - 1)
+                        End = blocks - 1
                     };
                     sidecar.BlockMedia[0].DumpHardwareArray[0].Manufacturer = dev.Manufacturer;
                     sidecar.BlockMedia[0].DumpHardwareArray[0].Model = dev.Model;
